requirements.txt
websockets
requests
better_profanity
openai
python-dotenv
google-auth